+++ changelog.md
@@ -1,5 +1,14 @@
 # pyxiv-dl Changelog
 
+### 0.3.0
+
+* Added download range option (`-r`/`--range`) for multiple image posts
+    * This enables you to download only a specific range of images based from the given starting and ending ranges.
+* Added download index option (`-i`/`--index`) for multiple image posts
+    * If you want to download only the specific image based on its image position (index) in the post, you can use this option.
+* Necessary regression: multiple post download (e.g. `post1 post2 postn ...` is no longer possible due to the implementation of download range and download index options)
+* Some bugfixes here and there, including some code smells reported by SonarCloud
+
 ### 0.2.0
 
 * Added ugoira support